Cover currency overlay supportability directly
